Data type size

+++ 0x00-hello_world/6-size.c
@@ -13,10 +13,10 @@ long int long_int_size;
 long long int long_long_int_size;
 float float_size;
 
-printf("Size of a char: %lu byte(s)\n", sizeof(charsize));
-printf("Size of int: %lu byte(s)\n", sizeof(intvalue));
-printf("Size of long int: %lu byte(s)\n", sizeof(long_int_size));
-printf("Size of long long int: %lu byte(s)\n", sizeof(long_long_int_size));
-printf("Size of float : %lu byte(s)\n", sizeof(float));
+printf("Size of a char: %lu byte(s)\n", (unsigned long) sizeof(charsize));
+printf("Size of int: %lu byte(s)\n", (unsigned long) sizeof(intvalue));
+printf("Size of long int: %lu byte(s)\n", (unsigned long)  sizeof(long_int_size));
+printf("Size of long long int: %lu byte(s)\n", (unsigned long) sizeof(long_long_int_size));
+printf("Size of float : %lu byte(s)\n", (unsigned long) sizeof(float));
 return (0);
 }
